# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.CouponUsageResponse:
    properties:
      coupon:
        allOf:
        - $ref: '#/definitions/model.CouponResponse'
        description: Related data
      coupon_id:
        description: Coupon ID
        example: 1
        type: integer
      created_at:
        description: Creation time
        example: "2024-01-01T00:00:00Z"
        type: string
      currency:
        description: Currency
        example: USD
        type: string
      discount_amount:
        description: Discount amount
        example: 5.99
        type: number
      id:
        description: Usage ID
        example: 1
        type: integer
      order_amount:
        description: Original order amount
        example: 29.99
        type: number
      subscription_order:
        allOf:
        - $ref: '#/definitions/model.SubscriptionOrderResponse'
        description: Order info
      subscription_order_id:
        description: Order ID
        example: 1
        type: integer
      updated_at:
        description: Update time
        example: "2024-01-01T00:00:00Z"
        type: string
      user:
        allOf:
        - $ref: '#/definitions/model.UserResponse'
        description: User info
      user_id:
        description: User ID
        example: 1
        type: integer
    type: object
...
  model.PaymentRecordResponse:
    properties:
      amount:
        description: Payment amount
        example: 29.99
        type: number
      can_refund:
        description: Refundable status
        type: boolean
      created_at:
        description: Creation time
        example: "2024-01-01T00:00:00Z"
        type: string
      currency:
        description: Currency
        example: CNY
        type: string
      exchange_rate:
        description: Exchange rate
        example: 1
        type: number
      expired_at:
        description: Expiration time
        example: "2024-01-01T01:00:00Z"
        type: string
      gateway:
        description: Payment gateway
        example: epay
        type: string
      id:
        description: Payment ID
        example: 1
        type: integer
      is_expired:
        description: Computed fields
        type: boolean
      notified_at:
        description: Notification time
        example: "2024-01-01T00:31:00Z"
        type: string
      out_trade_no:
        description: Merchant order number
        example: ORDER202401010001
        type: string
      paid_at:
        description: Payment completion time
        example: "2024-01-01T00:30:00Z"
        type: string
      payment_method:
        description: Payment method
        example: alipay
        type: string
      payment_no:
        description: Payment number
        example: PAY202401010001
        type: string
      payment_status:
        description: Gateway payment status
        example: success
        type: string
      payment_url:
        description: Payment URL
        example: https://example.com/pay
        type: string
      qr_code_url:
        description: QR code URL
        example: https://example.com/qr
        type: string
      refund_amount:
        description: Refund amount
        example: 0
        type: number
      refund_reason:
        description: Refund reason
        example: User request
        type: string
      refund_status:
        description: Refund status
        example: none
        type: string
      refundable_amount:
        description: Refundable amount
        type: number
      refunded_at:
        description: Refund time
        example: "2024-01-02T10:00:00Z"
        type: string
      remark:
        description: Remark
        example: Subscription payment
        type: string
      status:
        description: Payment status
        example: completed
        type: string
      subscription_order:
        allOf:
        - $ref: '#/definitions/model.SubscriptionOrderResponse'
        description: Subscription order info
      subscription_order_id:
        description: Subscription order ID
        example: 1
        type: integer
      transaction_id:
        description: Transaction ID
        example: TXN123456789
        type: string
      updated_at:
        description: Update time
        example: "2024-01-01T00:30:00Z"
        type: string
      user:
        allOf:
        - $ref: '#/definitions/model.UserResponse'
        description: Related data
      user_id:
        description: User ID
        example: 1
        type: integer
    type: object
...
  model.SubscriptionOrderResponse:
    properties:
      amount:
        description: Amount
        example: 29.99
        type: number
      billing_period_end:
        description: Billing end
        example: "2024-02-01T00:00:00Z"
        type: string
      billing_period_start:
        description: Billing start
        example: "2024-01-01T00:00:00Z"
        type: string
      coupon_code:
        description: Coupon code
        example: SAVE20
        type: string
      created_at:
        description: Creation time
        example: "2024-01-01T00:00:00Z"
        type: string
      currency:
        description: Currency
        example: USD
        type: string
      discount_amount:
        description: Discount amount
        example: 0
        type: number
      discount_type:
        description: Discount type
        example: percentage
        type: string
      discount_value:
        description: Discount value
        example: 20
        type: number
      id:
        description: Order ID
        example: 1
        type: integer
      invoice_number:
        description: Invoice number
        example: INV-2024-001
        type: string
      invoice_status:
        description: Invoice status
        example: sent
        type: string
      invoiced_at:
        description: Invoiced time
        example: "2024-01-01T10:30:00Z"
        type: string
      order_number:
        description: Order number
        example: ORD-2024-001
        type: string
      order_type:
        description: Order type
        example: new
        type: string
      paid_at:
        description: Paid time
        example: "2024-01-01T10:30:00Z"
        type: string
      payment_gateway:
        description: Payment gateway
        example: stripe
        type: string
      payment_method:
        description: Payment method
        example: credit_card
        type: string
      payment_status:
        description: Payment status
        example: completed
        type: string
      refund_amount:
        description: Refund amount
        example: 0
        type: number
      refund_reason:
        description: Refund reason
        example: User request
        type: string
      refunded_at:
        description: Refunded time
        example: "2024-01-15T10:30:00Z"
        type: string
      setup_fee:
        description: Setup fee
        example: 0
        type: number
      status:
        description: Status
        example: paid
        type: string
      subscription_plan:
        allOf:
        - $ref: '#/definitions/model.SubscriptionPlanResponse'
        description: Plan info
      subscription_plan_id:
        description: Plan ID
        example: 1
        type: integer
      total_amount:
        description: Total amount
        example: 29.99
        type: number
      transaction_id:
        description: Transaction ID
        example: txn_123456
        type: string
      updated_at:
        description: Update time
        example: "2024-01-01T00:00:00Z"
        type: string
      user:
        allOf:
        - $ref: '#/definitions/model.UserResponse'
        description: Related data
      user_id:
        description: User ID
        example: 1
        type: integer
      user_subscription:
        allOf:
        - $ref: '#/definitions/model.UserSubscriptionResponse'
        description: Subscription info
      user_subscription_id:
        description: Subscription ID
        example: 1
        type: integer
    type: object
...
  model.UserResponse:
    properties:
      avatar:
        type: string
      created_at:
        description: Timestamp Fields
        type: string
      deleted_at:
        type: string
      email:
        description: Core Identity Fields
        type: string
      github_id:
        type: string
      google_id:
        description: OAuth Provider IDs (only show if not empty)
        type: string
      id:
        description: Primary Key
        type: integer
      invite_code_id:
        description: Invite Code Fields
        type: integer
      invite_code_used:
        type: string
      name:
        type: string
      provider:
        description: Authentication Fields (excluding password)
        type: string
      provider_data:
        description: Provider Metadata (only show if not empty)
        type: string
      role:
        type: string
      status:
        type: string
      telegram_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  response.BadRequestResponse:
    description: Bad Request response format
    properties:
      code:
        example: 4000
        type: integer
      message:
        example: Invalid request parameters
        type: string
    type: object
...
  response.ForbiddenResponse:
    description: Forbidden response format
    properties:
      code:
        example: 4003
        type: integer
      message:
        example: Access denied
        type: string
    type: object
  response.InternalServerErrorResponse:
    description: Internal Server Error response format
    properties:
      code:
        example: 5000
        type: integer
      message:
        example: Internal server error
        type: string
    type: object
...
  response.NotFoundResponse:
    description: Not Found response format
    properties:
      code:
        example: 4004
        type: integer
      message:
        example: Resource not found
        type: string
    type: object
  response.PaginatedResponse:
    description: Paginated response format
    properties:
      code:
        example: 0
        type: integer
      data:
        items:
          type: object
        type: array
      limit:
        example: 10
        type: integer
      message:
        example: success
        type: string
      offset:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  response.StandardResponse:
    description: Standard API response format
    properties:
      code:
        description: Response code (0 for success, non-zero for errors)
        example: 0
        type: integer
      data:
        description: Response data (optional)
        type: object
      message:
        description: Response message
        example: success
        type: string
    type: object
  response.UnauthorizedResponse:
    description: Unauthorized response format
    properties:
      code:
        example: 4001
        type: integer
      message:
        example: Authentication required
        type: string
    type: object
...
  service.CreateSubscriptionOrderRequest:
    properties:
      coupon_code:
        example: SAVE20
        type: string
      metadata:
        type: string
      order_type:
        enum:
        - new
        - renewal
        - upgrade
        - downgrade
        example: new
        type: string
      payment_gateway:
        example: epay
        type: string
      payment_method:
        example: alipay
        type: string
      return_url:
        example: https://example.com/payment/return
        type: string
      subscription_plan_id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    required:
    - order_type
    - payment_gateway
    - payment_method
    - subscription_plan_id
    - user_id
    type: object
  service.CreateSubscriptionOrderResponse:
    properties:
      expired_at:
        type: string
      order:
        $ref: '#/definitions/model.SubscriptionOrderResponse'
      payment_record:
        $ref: '#/definitions/model.PaymentRecordResponse'
      payment_url:
        type: string
      qr_code_url:
        type: string
    type: object
...
  service.ValidateCouponResponse:
    properties:
      coupon:
        $ref: '#/definitions/model.CouponResponse'
      discount_amount:
        example: 5.99
        type: number
      final_amount:
        example: 24
        type: number
      message:
        example: Coupon is valid
        type: string
      valid:
        example: true
        type: boolean
    type: object
...
  user.ValidateCouponRequest:
    properties:
      coupon_code:
        example: SAVE20
        type: string
      order_amount:
        example: 29.99
        minimum: 0
        type: number
      plan_id:
        example: 1
        type: integer
    required:
    - coupon_code
    - order_amount
    - plan_id
    type: object
host: localhost:8080
info:
  contact: {}
  description: A comprehensive service management platform with subscription-based
    billing, user management, and server administration. Features include OAuth2 authentication,
    traffic subscription management, multi-gateway payments, referral programs, and
    customer support system.
  title: Linke API
  version: "1.0"
paths:
...
  /subscription-orders:
    post:
      consumes:
      - application/json
      description: Create a new subscription order with payment
      parameters:
      - description: Subscription order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/service.CreateSubscriptionOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.CreateSubscriptionOrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[User] Create subscription order'
      tags:
      - User-Subscription
  /subscription-orders/{id}:
    get:
      consumes:
      - application/json
      description: Get subscription order details
      parameters:
      - description: Subscription order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.SubscriptionOrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ForbiddenResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[User] Get subscription order'
      tags:
      - User-Subscription
  /subscription-orders/my:
    get:
      consumes:
      - application/json
      description: Get current user's subscription orders
      parameters:
      - description: Limit results
        example: 10
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Offset results
        example: 0
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.SubscriptionOrderResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[User] Get my subscription orders'
      tags:
      - User-Subscription
...
  /tasks:
    post:
      consumes:
      - application/json
      description: Create and enqueue a new task
      parameters:
      - description: Task details
        in: body
        name: task
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - Task-System
  /tasks/status:
    get:
      description: Get the current status of the task queue
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: Get queue status
      tags:
      - Task-System
  /user/coupons/usages:
    get:
      consumes:
      - application/json
      description: Get a paginated list of current user's coupon usage records
      parameters:
      - description: Number of items per page (1-100)
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Number of items to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.CouponUsageResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: '[User] Get my coupon usage history'
      tags:
      - User - Coupons
  /user/coupons/validate:
    post:
      consumes:
      - application/json
      description: Validate a coupon code for a specific plan and order amount
      parameters:
      - description: Coupon validation data
        in: body
        name: validate_request
        required: true
        schema:
          $ref: '#/definitions/user.ValidateCouponRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.ValidateCouponResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[User] Validate coupon code'
      tags:
      - User-Subscription
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"